refactor based on Stefan's suggestion
