+++ src/zenml/cli/__init__.py
@@ -812,11 +812,15 @@ zenml model-deployer logs MODEL_DEPLOYER_NAME
 
 """
 
+from zenml.cli.annotator import *  # noqa
 from zenml.cli.base import *  # noqa
 from zenml.cli.config import *  # noqa
 from zenml.cli.example import *  # noqa
+from zenml.cli.feature import *  # noqa
 from zenml.cli.integration import *  # noqa
+from zenml.cli.model import *  # noqa
 from zenml.cli.pipeline import *  # noqa
+from zenml.cli.secret import *  # noqa
 from zenml.cli.server import *  # noqa
 from zenml.cli.stack import *  # noqa
 from zenml.cli.stack_components import *  # noqa
