+++ src/zenml/cli/stack_components.py
@@ -13,9 +13,7 @@
 #  permissions and limitations under the License.
 """Functionality to generate stack component CLI commands."""
 
-import getpass
 import time
-import uuid
 from importlib import import_module
 from typing import (
     TYPE_CHECKING,
@@ -26,7 +24,6 @@ from typing import (
     Sequence,
     Tuple,
     Type,
-    cast,
 )
 
 import click
@@ -34,37 +31,25 @@ from pydantic import ValidationError
 from rich.markdown import Markdown
 
 from zenml.cli import utils as cli_utils
+from zenml.cli.annotator import register_annotator_subcommands
 from zenml.cli.cli import TagGroup, cli
-from zenml.cli.utils import (
-    confirmation,
-    declare,
-    error,
-    expand_argument_value_from_file,
-    parse_unknown_options,
-    pretty_print_model_deployer,
-    pretty_print_secret,
-    print_list_items,
-    print_served_model_configuration,
-    warning,
-)
+from zenml.cli.feature import register_feature_store_subcommands
+from zenml.cli.model import register_model_deployer_subcommands
+from zenml.cli.secret import register_secrets_manager_subcommands
 from zenml.config.global_config import GlobalConfiguration
 from zenml.console import console
 from zenml.constants import MANDATORY_COMPONENT_ATTRIBUTES
 from zenml.enums import CliCategories, StackComponentType
-from zenml.exceptions import EntityExistsError, SecretExistsError
+from zenml.exceptions import EntityExistsError
 from zenml.io import fileio
 from zenml.repository import Repository
-from zenml.secret import ARBITRARY_SECRET_SCHEMA_TYPE
 from zenml.stack import StackComponent
 from zenml.utils.analytics_utils import AnalyticsEvent, track_event
 from zenml.utils.source_utils import validate_flavor_source
 from zenml.zen_stores.models.component_wrapper import ComponentWrapper
 
 if TYPE_CHECKING:
-    from zenml.annotators.base_annotator import BaseAnnotator
-    from zenml.feature_stores.base_feature_store import BaseFeatureStore
-    from zenml.model_deployers import BaseModelDeployer
-    from zenml.secrets_managers.base_secrets_manager import BaseSecretsManager
+    pass
 
 
 def _get_required_attributes(
@@ -1378,1128 +1363,6 @@ def register_single_stack_component_cli_commands(
     command_group.command(
         "explain", help=f"Explaining the {plural_display_name}."
     )(explain_command)
-
-
-def register_secrets_manager_subcommands() -> None:
-    """Registers CLI subcommands for the Secrets Manager."""
-    secrets_manager_group = cast(TagGroup, cli.commands.get("secrets-manager"))
-    if secrets_manager_group:
-
-        @secrets_manager_group.group(
-            cls=TagGroup,
-            help="Commands for interacting with secrets.",
-        )
-        @click.pass_context
-        def secret(ctx: click.Context) -> None:
-            """List and manage your secrets.
-
-            Args:
-                ctx: Click context.
-            """
-            repo = Repository()
-            active_stack = repo.zen_store.get_stack(name=repo.active_stack_name)
-            secrets_manager_wrapper = active_stack.get_component_wrapper(
-                StackComponentType.SECRETS_MANAGER
-            )
-            if secrets_manager_wrapper is None:
-                error(
-                    "No active secrets manager found. Please create a secrets manager "
-                    "first and add it to your stack."
-                )
-                return
-
-            ctx.obj = secrets_manager_wrapper.to_component()
-
-        @secret.command(
-            "register",
-            context_settings={"ignore_unknown_options": True},
-            help="Register a secret with the given name and schema.",
-        )
-        @click.argument("name", type=click.STRING)
-        @click.option(
-            "--schema",
-            "-s",
-            "secret_schema_type",
-            default=ARBITRARY_SECRET_SCHEMA_TYPE,
-            help="Register a secret with an optional schema.",
-            type=str,
-        )
-        @click.option(
-            "--interactive",
-            "-i",
-            "interactive",
-            is_flag=True,
-            help="Use interactive mode to enter the secret values.",
-            type=click.BOOL,
-        )
-        @click.argument("args", nargs=-1, type=click.UNPROCESSED)
-        @click.pass_obj
-        def register_secret(
-            secrets_manager: "BaseSecretsManager",
-            name: str,
-            secret_schema_type: str,
-            interactive: bool,
-            args: List[str],
-        ) -> None:
-            """Register a secret with the given name and schema.
-
-            Use this command to store sensitive information into a ZenML secret. The
-            secret data consists of key-value pairs that can be configured interactively
-            (if the `--interactive` option is set) or via command-line arguments.
-            If a schema is indicated, the secret key-value pairs will be validated
-            against the schema.
-
-            When passed as command line arguments, the secret field values may also be
-            loaded from files instead of being issued inline, by prepending the field
-            name with a `@` sign. For example, the following command line:
-                zenml secrets-manager secret register my_secret --secret_token=@/path/to/file.json
-            will load the value for the field `secret_token` from the file
-            `/path/to/file.json`.
-            To use the `@` sign as the first character of a field name without pointing
-            to a file, double the `@` sign. For example, the following command line:
-                zenml secrets-manager secret register my_secret --username=zenml --password=@@password
-            will interpret the value of the field `password` as the literal string
-            `@password`.
-
-            Examples:
-            - register a secret with the name `secret_one` and configure its values
-            interactively:
-                zenml secrets-manager secret -manager secret register secret_one -i
-            - register a secret with the name `secret_two` and configure its values
-            via command line arguments:
-                zenml secrets-manager secret register secret_two --username=admin --password=secret
-            - register a secret with the name `secret_three` interactively and
-            conforming to a schema named `aws` (which is defined in the `aws`
-            integration):
-                zenml integration install aws
-                zenml secrets-manager secret register secret_three -i --schema=aws
-            - register a secret with the name `secret_four` from command line arguments
-            and conforming to a schema named `aws` (which is defined in the `aws`
-            integration). Also load the value for the field `secret_token` from a
-            local file:
-                zenml integration install aws
-                zenml secrets-manager secret register secret_four --schema=aws \
-                    --aws_access_key_id=1234567890 \
-                    --aws_secret_access_key=abcdefghij \
-                    --aws_session_token=@/path/to/token.txt
-
-
-            Args:
-                secrets_manager: The secrets manager to use.
-                name: The name of the secret to register.
-                secret_schema_type: The schema to use for validation.
-                interactive: Whether to use interactive mode to enter the secret values.
-                args: Command line arguments.
-            """
-            # flake8: noqa: C901
-
-            # TODO [ENG-871]: Formatting for `zenml secrets-manager secret register --help` currently
-            #  broken.
-            # TODO [ENG-725]: Allow passing in json/dict when registering a secret as an
-            #   additional option for the user on top of the interactive
-            try:
-                parsed_args = parse_unknown_options(args, expand_args=True)
-            except AssertionError as e:
-                error(str(e))
-
-            if "name" in parsed_args:
-                error(
-                    "You can't use 'name' as the key for one of your secrets."
-                )
-            elif name == "name":
-                error("Secret names cannot be named 'name'.")
-
-            if name.startswith("--"):
-                error(
-                    "Secret names cannot start with '--' The first argument must be."
-                    "the secret name."
-                )
-
-            if "name" in parsed_args:
-                error("Secret names cannot be passed as arguments.")
-
-            try:
-                from zenml.secret.secret_schema_class_registry import (
-                    SecretSchemaClassRegistry,
-                )
-
-                secret_schema = SecretSchemaClassRegistry.get_class(
-                    secret_schema=secret_schema_type
-                )
-            except KeyError as e:
-                error(str(e))
-
-            secret_keys = secret_schema.get_schema_keys()
-
-            secret_contents = {"name": name}
-
-            if interactive:
-
-                if parsed_args:
-                    error(
-                        "Cannot pass secret fields as arguments when using interactive "
-                        "mode."
-                    )
-
-                if secret_schema_type != ARBITRARY_SECRET_SCHEMA_TYPE:
-                    click.echo(
-                        "You have supplied a secret schema with predefined keys. "
-                        "You can fill these out sequentially now. Just press ENTER to "
-                        "skip optional secrets that you do not want to set"
-                    )
-                    for k in secret_keys:
-                        v = getpass.getpass(f"Secret value for {k}:")
-                        if v:
-                            secret_contents[
-                                k
-                            ] = expand_argument_value_from_file(name=k, value=v)
-                else:
-                    click.echo(
-                        "You have not supplied a secret schema with any "
-                        "predefined keys. Entering interactive mode:"
-                    )
-                    while True:
-                        k = click.prompt("Please enter a secret key")
-                        if k not in secret_contents:
-                            v = getpass.getpass(
-                                f"Please enter the secret value for the key [{k}]:"
-                            )
-                            secret_contents[
-                                k
-                            ] = expand_argument_value_from_file(name=k, value=v)
-                        else:
-                            warning(
-                                f"Key {k} already in this secret. Please restart this "
-                                f"process or use 'zenml secrets-manager secret update {name} --{k}=...' "
-                                f"to update this key after the secret is registered. "
-                                f"Skipping ..."
-                            )
-
-                        if not click.confirm(
-                            "Do you want to add another key-value pair to this secret?"
-                        ):
-                            break
-
-            else:
-                if not parsed_args:
-                    error(
-                        "Secret fields must be passed as arguments when not using "
-                        "interactive mode."
-                    )
-
-                secret_contents.update(parsed_args)
-
-            try:
-                secret = secret_schema(**secret_contents)
-            except ValidationError as e:
-                error(
-                    f"Secret values do not conform with the secret schema: {str(e)}"
-                )
-
-            click.echo("The following secret will be registered.")
-            pretty_print_secret(secret=secret, hide_secret=True)
-
-            with console.status(f"Saving secret `{name}`..."):
-                try:
-                    secrets_manager.register_secret(secret=secret)
-                except SecretExistsError:
-                    error(f"A secret with name '{name}' already exists.")
-
-        @secret.command("get", help="Get a secret, given its name.")
-        @click.argument("name", type=click.STRING)
-        @click.pass_obj
-        def get_secret(
-            secrets_manager: "BaseSecretsManager",
-            name: str,
-        ) -> None:
-            """Get a secret, given its name.
-
-            Args:
-                secrets_manager: The secrets manager to use.
-                name: The name of the secret to get.
-            """
-            try:
-                secret = secrets_manager.get_secret(secret_name=name)
-                pretty_print_secret(secret, hide_secret=False)
-            except KeyError as e:
-                error(
-                    f"Secret with name `{name}` does not exist or could not be loaded: "
-                    f"{str(e)}."
-                )
-
-        @secret.command(
-            "list", help="List all secrets tracked by your Secrets Manager."
-        )
-        @click.pass_obj
-        def list_secret(secrets_manager: "BaseSecretsManager") -> None:
-            """List all secrets tracked by your Secrets Manager.
-
-            Args:
-                secrets_manager: The secrets manager to use.
-            """
-            with console.status("Getting secret names..."):
-                secret_names = secrets_manager.get_all_secret_keys()
-                print_list_items(
-                    list_items=secret_names, column_title="SECRET_NAMES"
-                )
-
-        @secret.command(
-            "update",
-            context_settings={"ignore_unknown_options": True},
-            help="Update a secret with a given name.",
-        )
-        @click.argument("name", type=click.STRING)
-        @click.option(
-            "--interactive",
-            "-i",
-            "interactive",
-            is_flag=True,
-            help="Use interactive mode to update the secret values.",
-            type=click.BOOL,
-        )
-        @click.argument("args", nargs=-1, type=click.UNPROCESSED)
-        @click.pass_obj
-        def update_secret(
-            secrets_manager: "BaseSecretsManager",
-            name: str,
-            interactive: bool,
-            args: List[str],
-        ) -> None:
-            """Update a secret with a given name.
-
-            Use this command to update the information stored in an existing ZenML
-            secret. The secret's key-value pairs can be updated interactively
-            (if the `--interactive` option is set) or via command-line arguments.
-            If a schema is associated with the existing secret, the updated secret
-            key-value pairs will be validated against the schema.
-
-            When passed as command line arguments, the secret field values may also be
-            loaded from files instead of being issued inline, by prepending the field
-            name with a `@` sign. For example, the following command line:
-                zenml secrets-manager secret update my_secret --secret_token=@/path/to/file.json
-            will load the value for the field `secret_token` from the file
-            `/path/to/file.json`.
-            To use the `@` sign as the first character of a field name without pointing
-            to a file, double the `@` sign. For example, the following command line:
-                zenml secrets-manager secret update my_secret --username=zenml --password=@@password
-            will interpret the value of the field `password` as the literal string
-            `@password`.
-            
-            Examples:
-            - update a secret with the name `secret_one` and configure its values
-            interactively:
-                zenml secrets-manager secret update secret_one -i
-            - update a secret with the name `secret_two` from command line arguments
-            and load the value for the field `secret_token` from a
-            local file:
-                zenml secrets-manager secret update secret_four \
-                    --aws_access_key_id=1234567890 \
-                    --aws_secret_access_key=abcdefghij \
-                    --aws_session_token=@/path/to/token.txt
-
-            Args:
-                secrets_manager: The secrets manager to use.
-                name: The name of the secret to update.
-                interactive: Use interactive mode to update the secret values.
-                args: The command line arguments to use to update the secret.
-            """
-            # TODO [ENG-726]: allow users to pass in dict or json
-
-            with console.status(f"Getting secret `{name}`..."):
-                try:
-                    secret = secrets_manager.get_secret(secret_name=name)
-                except KeyError as e:
-                    error(
-                        f"Secret with name `{name}` does not exist or could not be "
-                        f"loaded: {str(e)}."
-                    )
-
-            try:
-                parsed_args = parse_unknown_options(args, expand_args=True)
-            except AssertionError as e:
-                error(str(e))
-                return
-
-            if "name" in parsed_args:
-                error("Secret names cannot be passed as arguments.")
-
-            updated_contents = {"name": name}
-
-            if interactive:
-                if parsed_args:
-                    error(
-                        "Cannot pass secret fields as arguments when using interactive "
-                        "mode."
-                    )
-
-                click.echo(
-                    "You will now have a chance to overwrite each secret "
-                    "one by one. Press enter to skip."
-                )
-                for key, value in secret.content.items():
-                    new_value = getpass.getpass(f"New value for " f"{key}:")
-                    if new_value:
-                        updated_contents[key] = new_value
-                    else:
-                        updated_contents[key] = value
-
-            else:
-
-                if not parsed_args:
-                    error(
-                        "Secret fields must be passed as arguments when not using "
-                        "interactive mode."
-                    )
-                updated_contents.update(secret.content)
-                updated_contents.update(parsed_args)
-
-            try:
-                updated_secret = secret.__class__(**updated_contents)
-            except ValidationError as e:
-                error(
-                    f"Secret values do not conform with the secret schema: {str(e)}"
-                )
-
-            click.echo("The following secret will be updated.")
-            pretty_print_secret(updated_secret, hide_secret=True)
-
-            with console.status(f"Updating secret `{name}`..."):
-                try:
-                    secrets_manager.update_secret(secret=updated_secret)
-                    console.print(f"Secret with name '{name}' has been updated")
-                except KeyError:
-                    error(f"Secret with name `{name}` already exists.")
-
-        @secret.command(
-            "delete", help="Delete a secret identified by its name."
-        )
-        @click.argument("name", type=click.STRING)
-        @click.option(
-            "--yes",
-            "-y",
-            type=click.BOOL,
-            default=False,
-            is_flag=True,
-            help="Skip asking for confirmation.",
-        )
-        @click.pass_obj
-        def delete_secret(
-            secrets_manager: "BaseSecretsManager",
-            name: str,
-            yes: bool = False,
-        ) -> None:
-            """Delete a secret identified by its name.
-
-            Args:
-                secrets_manager: The secrets manager to use.
-                name: The name of the secret to delete.
-                yes: Skip asking for confirmation.
-            """
-            if not yes:
-                confirmation_response = confirmation(
-                    f"This will delete all data associated with the `{name}` secret. "
-                    "Are you sure you want to proceed?"
-                )
-                if not confirmation_response:
-                    console.print("Aborting secret deletion...")
-                    return
-
-            with console.status(f"Deleting secret `{name}`..."):
-                try:
-                    secrets_manager.delete_secret(name)
-                    console.print(f"Deleted secret `{name}`.")
-                except KeyError:
-                    error(f"Secret with name `{name}` no longer present.")
-
-        @secret.command(
-            "cleanup",
-            hidden=True,
-            help="Delete all secrets tracked by your Secrets Manager.",
-        )
-        @click.option(
-            "--yes",
-            "-y",
-            "yes",
-            is_flag=True,
-            help="Force the deletion of all secrets",
-            type=click.BOOL,
-        )
-        @click.option(
-            "--force",
-            "-f",
-            "force",
-            is_flag=True,
-            help="DEPRECATED: Force the deletion of all secrets. Use `-y/--yes` "
-            "instead.",
-            type=click.BOOL,
-        )
-        @click.pass_obj
-        def delete_all_secrets(
-            secrets_manager: "BaseSecretsManager", yes: bool, force: bool
-        ) -> None:
-            """Delete all secrets tracked by your Secrets Manager.
-
-            Args:
-                secrets_manager: The secrets manager to use.
-                yes: Skip asking for confirmation.
-                force: DEPRECATED: Skip asking for confirmation.
-            """
-            if force:
-                warning(
-                    "The `--force` flag will soon be deprecated. Use `--yes` or `-y` "
-                    "instead."
-                )
-            if not yes:
-                confirmation_response = confirmation(
-                    "This will delete all secrets. Are you sure you want to proceed?"
-                )
-                if not confirmation_response:
-                    console.print("Aborting deletion of all secrets...")
-                    return
-
-            with console.status("Deleting all secrets..."):
-                secrets_manager.delete_all_secrets()
-                console.print("Deleted all secrets.")
-
-
-def register_feature_store_subcommands() -> None:
-    """Registers CLI subcommands for the Feature Store."""
-    feature_store_group = cast(TagGroup, cli.commands.get("feature-store"))
-    if feature_store_group:
-
-        @feature_store_group.group(
-            cls=TagGroup,
-            help="Commands for interacting with your features.",
-        )
-        @click.pass_context
-        def feature(ctx: click.Context) -> None:
-            """Features as obtained from a feature store.
-
-            Args:
-                ctx: The click context.
-            """
-            repo = Repository()
-            active_stack = repo.zen_store.get_stack(name=repo.active_stack_name)
-            feature_store_wrapper = active_stack.get_component_wrapper(
-                StackComponentType.FEATURE_STORE
-            )
-            if feature_store_wrapper is None:
-                error(
-                    "No active feature store found. Please create a feature store "
-                    "first and add it to your stack."
-                )
-                return
-            ctx.obj = feature_store_wrapper.to_component()
-
-        @feature.command("get-data-sources")
-        @click.pass_obj
-        def get_data_sources(feature_store: "BaseFeatureStore") -> None:
-            """Get all data sources from the feature store.
-
-            Args:
-                feature_store: The feature store.
-            """
-            data_sources = feature_store.get_data_sources()
-            declare(f"Data sources: {data_sources}")
-
-        @feature.command("get-entities")
-        @click.pass_obj
-        def get_entities(feature_store: "BaseFeatureStore") -> None:
-            """Get all entities from the feature store.
-
-            Args:
-                feature_store: The feature store.
-            """
-            entities = feature_store.get_entities()
-            declare(f"Entities: {entities}")
-
-        @feature.command("get-feature-services")
-        @click.pass_obj
-        def get_feature_services(feature_store: "BaseFeatureStore") -> None:
-            """Get all feature services from the feature store.
-
-            Args:
-                feature_store: The feature store.
-            """
-            feature_services = feature_store.get_feature_services()
-            declare(f"Feature services: {feature_services}")
-
-        @feature.command("get-feature-views")
-        @click.pass_obj
-        def get_feature_views(feature_store: "BaseFeatureStore") -> None:
-            """Get all feature views from the feature store.
-
-            Args:
-                feature_store: The feature store.
-            """
-            feature_views = feature_store.get_feature_views()
-            declare(f"Feature views: {feature_views}")
-
-        @feature.command("get-project")
-        @click.pass_obj
-        def get_project(feature_store: "BaseFeatureStore") -> None:
-            """Get the current project name from the feature store.
-
-            Args:
-                feature_store: The feature store.
-            """
-            project = feature_store.get_project()
-            declare(f"Project name: {project}")
-
-        @feature.command("get-feast-version")
-        @click.pass_obj
-        def get_feast_version(feature_store: "BaseFeatureStore") -> None:
-            """Get the current Feast version being used.
-
-            Args:
-                feature_store: The feature store.
-            """
-            version = feature_store.get_feast_version()
-            declare(f"Feast version: {version}")
-
-
-def register_model_deployer_subcommands() -> None:
-    """Registers CLI subcommands for the Model Deployer."""
-    model_deployer_group = cast(TagGroup, cli.commands.get("model-deployer"))
-    if model_deployer_group:
-
-        @model_deployer_group.group(
-            cls=TagGroup,
-            help="Commands for interacting with annotation datasets.",
-        )
-        @click.pass_context
-        def models(ctx: click.Context) -> None:
-            """List and manage served models with the active model deployer.
-
-            Args:
-                ctx: The click context.
-            """
-            repo = Repository()
-            active_stack = repo.zen_store.get_stack(name=repo.active_stack_name)
-            model_deployer_wrapper = active_stack.get_component_wrapper(
-                StackComponentType.MODEL_DEPLOYER
-            )
-            if model_deployer_wrapper is None:
-                error(
-                    "No active model deployer found. Please add a model_deployer to "
-                    "your stack."
-                )
-                return
-            ctx.obj = model_deployer_wrapper.to_component()
-
-        @models.command(
-            "list",
-            help="Get a list of all served models within the model-deployer stack component.",
-        )
-        @click.option(
-            "--pipeline",
-            "-p",
-            type=click.STRING,
-            default=None,
-            help="Show only served models that were deployed by the indicated "
-            "pipeline.",
-        )
-        @click.option(
-            "--step",
-            "-s",
-            type=click.STRING,
-            default=None,
-            help="Show only served models that were deployed by the indicated "
-            "pipeline step.",
-        )
-        @click.option(
-            "--pipeline-run",
-            "-r",
-            type=click.STRING,
-            default=None,
-            help="Show only served models that were deployed by the indicated "
-            "pipeline run.",
-        )
-        @click.option(
-            "--model",
-            "-m",
-            type=click.STRING,
-            default=None,
-            help="Show only served model versions for the given model name.",
-        )
-        @click.option(
-            "--running",
-            is_flag=True,
-            help="Show only model servers that are currently running.",
-        )
-        @click.pass_obj
-        def list_models(
-            model_deployer: "BaseModelDeployer",
-            pipeline: Optional[str],
-            step: Optional[str],
-            pipeline_run: Optional[str],
-            model: Optional[str],
-            running: bool,
-        ) -> None:
-            """Get a list of all served models within the model-deployer stack component.
-
-            Args:
-                model_deployer: The model-deployer stack component.
-                pipeline: Show only served models that were deployed by the indicated
-                    pipeline.
-                step: Show only served models that were deployed by the indicated
-                    pipeline step.
-                pipeline_run: Show only served models that were deployed by the
-                    indicated pipeline run.
-                model: Show only served model versions for the given model name.
-                running: Show only model servers that are currently running.
-            """
-            services = model_deployer.find_model_server(
-                running=running,
-                pipeline_name=pipeline,
-                pipeline_run_id=pipeline_run,
-                pipeline_step_name=step,
-                model_name=model,
-            )
-            if services:
-                pretty_print_model_deployer(
-                    services,
-                    model_deployer,
-                )
-            else:
-                warning("No served models found.")
-
-        @models.command("describe", help="Describe a specified served model.")
-        @click.argument("served_model_uuid", type=click.STRING)
-        @click.pass_obj
-        def describe_model(
-            model_deployer: "BaseModelDeployer", served_model_uuid: str
-        ) -> None:
-            """Describe a specified served model.
-
-            Args:
-                model_deployer: The model-deployer stack component.
-                served_model_uuid: The UUID of the served model.
-            """
-            served_models = model_deployer.find_model_server(
-                service_uuid=uuid.UUID(served_model_uuid)
-            )
-            if served_models:
-                print_served_model_configuration(
-                    served_models[0], model_deployer
-                )
-                return
-            warning(
-                f"No model with uuid: '{served_model_uuid}' could be found."
-            )
-            return
-
-        @models.command(
-            "get-url",
-            help="Return the prediction URL to a specified model server.",
-        )
-        @click.argument("served_model_uuid", type=click.STRING)
-        @click.pass_obj
-        def get_url(
-            model_deployer: "BaseModelDeployer", served_model_uuid: str
-        ) -> None:
-            """Return the prediction URL to a specified model server.
-
-            Args:
-                model_deployer: The model-deployer stack component.
-                served_model_uuid: The UUID of the served model.
-            """
-            served_models = model_deployer.find_model_server(
-                service_uuid=uuid.UUID(served_model_uuid)
-            )
-            if served_models:
-                try:
-                    prediction_url = model_deployer.get_model_server_info(
-                        served_models[0]
-                    ).get("PREDICTION_URL")
-                    prediction_hostname = (
-                        model_deployer.get_model_server_info(
-                            served_models[0]
-                        ).get("PREDICTION_HOSTNAME")
-                        or "No hostname specified for this service"
-                    )
-                    declare(
-                        f"  Prediction URL of Served Model {served_model_uuid} "
-                        f"is:\n"
-                        f"  {prediction_url}\n"
-                        f"  and the hostname is: {prediction_hostname}"
-                    )
-                except KeyError:
-                    warning(
-                        "The deployed model instance has no 'prediction_url'."
-                    )
-                return
-            warning(
-                f"No model with uuid: '{served_model_uuid}' could be found."
-            )
-            return
-
-        @models.command("start", help="Start a specified model server.")
-        @click.argument("served_model_uuid", type=click.STRING)
-        @click.option(
-            "--timeout",
-            "-t",
-            type=click.INT,
-            default=300,
-            help="Time in seconds to wait for the model to start. Set to 0 to return "
-            "immediately after telling the server to start, without waiting for it to "
-            "become fully active (default: 300s).",
-        )
-        @click.pass_obj
-        def start_model_service(
-            model_deployer: "BaseModelDeployer",
-            served_model_uuid: str,
-            timeout: int,
-        ) -> None:
-            """Start a specified model server.
-
-            Args:
-                model_deployer: The model-deployer stack component.
-                served_model_uuid: The UUID of the served model.
-                timeout: Time in seconds to wait for the model to start.
-            """
-            served_models = model_deployer.find_model_server(
-                service_uuid=uuid.UUID(served_model_uuid)
-            )
-            if served_models:
-                model_deployer.start_model_server(
-                    served_models[0].uuid, timeout=timeout
-                )
-                declare(f"Model server {served_models[0]} was started.")
-                return
-
-            warning(
-                f"No model with uuid: '{served_model_uuid}' could be found."
-            )
-            return
-
-        @models.command("stop", help="Stop a specified model server.")
-        @click.argument("served_model_uuid", type=click.STRING)
-        @click.option(
-            "--timeout",
-            "-t",
-            type=click.INT,
-            default=300,
-            help="Time in seconds to wait for the model to start. Set to 0 to return "
-            "immediately after telling the server to stop, without waiting for it to "
-            "become inactive (default: 300s).",
-        )
-        @click.option(
-            "--yes",
-            "-y",
-            "force",
-            is_flag=True,
-            help="Force the model server to stop. This will bypass any graceful "
-            "shutdown processes and try to force the model server to stop immediately, "
-            "if possible.",
-        )
-        @click.option(
-            "--force",
-            "-f",
-            "old_force",
-            is_flag=True,
-            help="DEPRECATED: Force the model server to stop. This will bypass any graceful "
-            "shutdown processes and try to force the model server to stop immediately, "
-            "if possible. Use `-y/--yes` instead.",
-        )
-        @click.pass_obj
-        def stop_model_service(
-            model_deployer: "BaseModelDeployer",
-            served_model_uuid: str,
-            timeout: int,
-            force: bool,
-            old_force: bool,
-        ) -> None:
-            """Stop a specified model server.
-
-            Args:
-                model_deployer: The model-deployer stack component.
-                served_model_uuid: The UUID of the served model.
-                timeout: Time in seconds to wait for the model to stop.
-                force: Force the model server to stop.
-                old_force: DEPRECATED: Force the model server to stop.
-            """
-            if old_force:
-                force = old_force
-                warning(
-                    "The `--force` flag will soon be deprecated. Use `--yes` or `-y` instead."
-                )
-            served_models = model_deployer.find_model_server(
-                service_uuid=uuid.UUID(served_model_uuid)
-            )
-            if served_models:
-                model_deployer.stop_model_server(
-                    served_models[0].uuid, timeout=timeout, force=force
-                )
-                declare(f"Model server {served_models[0]} was stopped.")
-                return
-
-            warning(
-                f"No model with uuid: '{served_model_uuid}' could be found."
-            )
-            return
-
-        @models.command("delete", help="Delete a specified model server.")
-        @click.argument("served_model_uuid", type=click.STRING)
-        @click.option(
-            "--timeout",
-            "-t",
-            type=click.INT,
-            default=300,
-            help="Time in seconds to wait for the model to be deleted. Set to 0 to "
-            "return immediately after stopping and deleting the model server, without "
-            "waiting for it to release all allocated resources.",
-        )
-        @click.option(
-            "--yes",
-            "-y",
-            "force",
-            is_flag=True,
-            help="Force the model server to stop and delete. This will bypass any "
-            "graceful shutdown processes and try to force the model server to stop and "
-            "delete immediately, if possible.",
-        )
-        @click.option(
-            "--force",
-            "-f",
-            "old_force",
-            is_flag=True,
-            help="DEPRECATED: Force the model server to stop and delete. This will bypass any "
-            "graceful shutdown processes and try to force the model server to stop and "
-            "delete immediately, if possible. Use `-y/--yes` instead.",
-        )
-        @click.pass_obj
-        def delete_model_service(
-            model_deployer: "BaseModelDeployer",
-            served_model_uuid: str,
-            timeout: int,
-            force: bool,
-            old_force: bool,
-        ) -> None:
-            """Delete a specified model server.
-
-            Args:
-                model_deployer: The model-deployer stack component.
-                served_model_uuid: The UUID of the served model.
-                timeout: Time in seconds to wait for the model to be deleted.
-                force: Force the model server to stop and delete.
-                old_force: DEPRECATED: Force the model server to stop and delete.
-            """
-            if old_force:
-                force = old_force
-                warning(
-                    "The `--force` flag will soon be deprecated. Use `--yes` or `-y` instead."
-                )
-            served_models = model_deployer.find_model_server(
-                service_uuid=uuid.UUID(served_model_uuid)
-            )
-            if served_models:
-                model_deployer.delete_model_server(
-                    served_models[0].uuid, timeout=timeout, force=force
-                )
-                declare(f"Model server {served_models[0]} was deleted.")
-                return
-
-            warning(
-                f"No model with uuid: '{served_model_uuid}' could be found."
-            )
-            return
-
-        @models.command("logs", help="Show the logs for a model server.")
-        @click.argument("served_model_uuid", type=click.STRING)
-        @click.option(
-            "--follow",
-            "-f",
-            is_flag=True,
-            help="Continue to output new log data as it becomes available.",
-        )
-        @click.option(
-            "--tail",
-            "-t",
-            type=click.INT,
-            default=None,
-            help="Only show the last NUM lines of log output.",
-        )
-        @click.option(
-            "--raw",
-            "-r",
-            is_flag=True,
-            help="Show raw log contents (don't pretty-print logs).",
-        )
-        @click.pass_obj
-        def get_model_service_logs(
-            model_deployer: "BaseModelDeployer",
-            served_model_uuid: str,
-            follow: bool,
-            tail: Optional[int],
-            raw: bool,
-        ) -> None:
-            """Display the logs for a model server.
-
-            Args:
-                model_deployer: The model-deployer stack component.
-                served_model_uuid: The UUID of the served model.
-                follow: Continue to output new log data as it becomes available.
-                tail: Only show the last NUM lines of log output.
-                raw: Show raw log contents (don't pretty-print logs).
-            """
-            served_models = model_deployer.find_model_server(
-                service_uuid=uuid.UUID(served_model_uuid)
-            )
-            if not served_models:
-                warning(
-                    f"No model with uuid: '{served_model_uuid}' could be found."
-                )
-                return
-
-            for line in model_deployer.get_model_server_logs(
-                served_models[0].uuid, follow=follow, tail=tail
-            ):
-                # don't pretty-print log lines that are already pretty-printed
-                if raw or line.startswith("\x1b["):
-                    print(line)
-                else:
-                    console.print(line)
-
-
-def register_annotator_subcommands() -> None:
-    """Registers CLI subcommands for the annotator."""
-    annotator_group = cast(TagGroup, cli.commands.get("annotator"))
-    if annotator_group:
-
-        @annotator_group.group(
-            cls=TagGroup,
-            help="Commands for interacting with annotation datasets.",
-        )
-        @click.pass_context
-        def dataset(ctx: click.Context) -> None:
-            """Interact with ZenML annotator datasets.
-
-            Args:
-                ctx: The click Context object.
-            """
-            repo = Repository()
-            active_stack = repo.zen_store.get_stack(name=repo.active_stack_name)
-            annotator_wrapper = active_stack.get_component_wrapper(
-                StackComponentType.ANNOTATOR
-            )
-            if annotator_wrapper is None:
-                cli_utils.error(
-                    "No active annotator found. Please register an annotator "
-                    "first and add it to your stack."
-                )
-                return
-
-            ctx.obj = annotator_wrapper.to_component()
-
-        @dataset.command(
-            "list",
-            help="List the available datasets.",
-        )
-        @click.pass_obj
-        def dataset_list(annotator: "BaseAnnotator") -> None:
-            """List the available datasets.
-
-            Args:
-                annotator: The annotator stack component.
-            """
-            cli_utils.print_list_items(
-                list_items=annotator.get_dataset_names(),
-                column_title="DATASETS",
-            )
-
-        @dataset.command("stats")
-        @click.argument("dataset_name", type=click.STRING)
-        @click.pass_obj
-        def dataset_stats(
-            annotator: "BaseAnnotator", dataset_name: str
-        ) -> None:
-            """Display statistics about a dataset.
-
-            Args:
-                annotator: The annotator stack component.
-                dataset_name: The name of the dataset.
-            """
-            try:
-                stats = annotator.get_dataset_stats(dataset_name)
-                labeled_task_count, unlabeled_task_count = stats
-            except IndexError:
-                cli_utils.error(
-                    f"Dataset {dataset_name} does not exist. Please use `zenml annotator dataset list` to "
-                    f"list the available datasets."
-                )
-                return
-
-            total_task_count = unlabeled_task_count + labeled_task_count
-            cli_utils.declare(
-                f"Annotation stats for '{dataset_name}' dataset:", bold=True
-            )
-            cli_utils.declare(f"Total annotation tasks: {total_task_count}")
-            cli_utils.declare(f"Labeled annotation tasks: {labeled_task_count}")
-            cli_utils.declare(
-                f"Unlabeled annotation tasks: {unlabeled_task_count}"
-            )
-
-        @dataset.command("delete")
-        @click.argument("dataset_name", type=click.STRING)
-        @click.option(
-            "--all",
-            "-a",
-            "all",
-            is_flag=True,
-            help="Use this flag to delete all datasets.",
-            type=click.BOOL,
-        )
-        @click.pass_obj
-        def dataset_delete(
-            annotator: "BaseAnnotator", dataset_name: str, all: bool
-        ) -> None:
-            """Delete a dataset.
-
-            If the --all flag is used, all datasets will be deleted.
-
-            Args:
-                annotator: The annotator stack component.
-                dataset_name: Name of the dataset to delete.
-                all: Whether to delete all datasets.
-            """
-            cli_utils.declare(f"Deleting your dataset '{dataset_name}'")
-            dataset_names = (
-                annotator.get_dataset_names() if all else [dataset_name]
-            )
-            for dataset_name in dataset_names:
-                annotator.delete_dataset(dataset_name=dataset_name)
-                cli_utils.declare(
-                    f"Dataset '{dataset_name}' has now been deleted."
-                )
-
-        @dataset.command(
-            "annotate", context_settings={"ignore_unknown_options": True}
-        )
-        @click.argument("dataset_name", type=click.STRING)
-        @click.pass_obj
-        def dataset_annotate(
-            annotator: "BaseAnnotator", dataset_name: str
-        ) -> None:
-            """Command to launch the annotation interface for a dataset.
-
-            Args:
-                annotator: The annotator stack component.
-                dataset_name: Name of the dataset
-
-            Raises:
-                ValueError: If the dataset does not exist.
-            """
-            cli_utils.declare(
-                f"Launching the annotation interface for dataset '{dataset_name}'."
-            )
-            try:
-                annotator.get_dataset(dataset_name=dataset_name)
-                annotator.launch(
-                    url=annotator.get_url_for_dataset(dataset_name)
-                )
-            except ValueError as e:
-                raise ValueError("Dataset does not exist.") from e
 
 
 def register_all_stack_component_cli_commands() -> None:
